prediction fruits/App_Mobile_AgriMind/requirements.txt
streamlit
pandas
numpy
joblib
torch
torchvision
Pillow
scikit-learn
ultralytics
opencv-python-headless
dill